Fixed delta frames in YUV2
git-svn-id: file:///Users/Eric/Developer/Other/YetiCodecSVN/trunk@69 a9bbc950-3a31-49f7-86fe-bb7552e34ce9

+++ yeti.h
@@ -10,7 +10,7 @@
 
 #include "resource.h"
 
-inline void* ALIGNED_MALLOC (void* ptr, int size, int align, char* str) 
+inline void* ALIGNED_MALLOC (void* ptr, size_t size, int align, char* str) 
 {
    if(ptr)
    {
